+++ examples/data-sources/morpheus_virtual_images/data-source.tf
@@ -1,0 +1,13 @@
+data "morpheus_virtual_images" "example_virtual_images" {
+  sort_ascending = true
+  source         = "Synced"
+  filter {
+    name   = "name"
+    values = ["Test*"]
+  }
+
+  filter {
+    name   = "type"
+    values = ["vmdk", "iso"]
+  }
+}